+++ Content.Server/IoC/ServerContentIoC.cs
@@ -26,7 +26,6 @@ using Content.Server.Worldgen.Tools;
 using Content.Shared.Administration.Logs;
 using Content.Shared.Administration.Managers;
 using Content.Shared.Kitchen;
-using Content.Server.Stories.Partners.Managers;
 
 namespace Content.Server.IoC
 {
@@ -39,7 +38,7 @@ namespace Content.Server.IoC
             IoCManager.Register<IMoMMILink, MoMMILink>();
             IoCManager.Register<IServerPreferencesManager, ServerPreferencesManager>();
             IoCManager.Register<IServerDbManager, ServerDbManager>();
-            IoCManager.Register<IPartnersManager, PartnersManager>(); // SPACE STORIES
+            IoCManager.Register<ISponsorDbManager, SponsorDbManager>();
             IoCManager.Register<RecipeManager, RecipeManager>();
             IoCManager.Register<INodeGroupFactory, NodeGroupFactory>();
             IoCManager.Register<IConnectionManager, ConnectionManager>();
